before a merge
